requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg[binary]==3.2.10
celery==5.5.3
redis==6.4.0
google-genai==1.39.1
Pillow==11.3.0
opencv-python-headless==4.12.0.88
numpy==2.2.6
gunicorn==23.0.0
python-json-logger==4.0.0
